pyproject.toml
[project]
name = "electricity-forecast"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.0.3",
    "google-cloud-storage>=2.17.0",
    "hydra-core==1.3.2",
    "ipykernel>=6.30.1",
    "matplotlib>=3.6.0",
    "mlflow>=3.0.0",
    "numpy>=1.26.4",
    "pandas>=1.5.0",
    "pydantic==2.11.5",
    "pytest>=8.4.1",
    "pyyaml>=6.0",
    "scikit-learn>=1.6.0",
    "statsmodels>=0.14.0",
    "uv-build>=0.8.12",
    "xgboost>=1.7.0",
]

[project.scripts]
electricity-forecast = "electricity_forecast:main"

[tool.setuptools.packages.find]
where = ["src"]

[build-system]
requires = ["uv_build>=0.8.9,<0.9.0"]
build-backend = "uv_build"
